Refactor structure list code
