+++ src/editor/layout/structure/list/EditorStructureList.tsx
@@ -3,20 +3,16 @@ import {useEditorContext, useEditorDispatcher} from "../../../reducer/EditorProv
 import {StructureTypes} from "../../../../data/model/structure/StructureTypes";
 import {UUID} from "../../../../data/model/UUID";
 import {StructureID, StructureModel} from "../../../../data/model/structure/StructureModel";
-import {BaseInputSlug} from "../../../component/slug/BaseInputSlug";
-import {BaseFolderSlug} from "../../../component/slug/BaseFolderSlug";
 import {SectionModel} from "../../../../data/model/structure/system/SectionModel";
-import {MinusIcon, QuestionMarkCircleIcon, StarIcon, SwatchIcon, TrashIcon} from "@heroicons/react/24/outline";
 import React from "react";
-import {BaseActionSlug} from "../../../component/slug/BaseActionSlug";
+import {EditorStructureListItem} from "./EditorStructureListItem";
 
 export function EditorStructureList() {
     const editor = useEditorContext()
     const dispatcher = useEditorDispatcher()
 
     const structureItem = React.useCallback(
-        (structure: StructureModel) => <StructureListItem {...structure}/>,
-        []
+        (structure: StructureModel) => <EditorStructureListItem {...structure}/>, []
     );
 
     if (!editor || !dispatcher) {
@@ -54,75 +50,4 @@ export function EditorStructureList() {
             onSelect={selectStructure}
         />
     );
-}
-
-function StructureListItem(structure: StructureModel) {
-    const editor = useEditorContext()
-    const dispatcher = useEditorDispatcher()
-    if (!editor || !dispatcher) {
-        return <></>;
-    }
-
-    const collapseSection = () => {
-        if (structure.type !== StructureTypes.Section) {
-            return;
-        }
-
-        dispatcher({
-            type: 'level_patch_structure',
-            payload: {
-                id: structure.id,
-                collapsed: !(structure as SectionModel).collapsed,
-            } as SectionModel,
-        })
-    }
-
-    const renameStructure = (name: string) => {
-        dispatcher({
-            type: 'level_patch_structure',
-            payload: {
-                id: structure.id,
-                name: name,
-            }
-        })
-    }
-
-    const removeStructure = () => {
-        dispatcher({
-            type: 'level_remove_structure',
-            payload: structure.id,
-        })
-    }
-
-    return <>
-        {structure.type === StructureTypes.Section && (
-            <BaseFolderSlug
-                collapsed={(structure as SectionModel).collapsed}
-                onCollapseToggle={collapseSection}
-            />
-        )}
-        <StructureListIcon {...structure} />
-        <BaseInputSlug
-            value={structure.name}
-            placeholder={structure.type}
-            onRename={renameStructure}
-        />
-        {editor.level.start === structure.id
-            ? <StarIcon className="w-8 p-2"/>
-            : <BaseActionSlug onClick={removeStructure}><TrashIcon className="w-4"/></BaseActionSlug>
-        }
-    </>
-}
-
-function StructureListIcon(structure: StructureModel) {
-    switch (structure.type) {
-        case StructureTypes.Chunk:
-            return <MinusIcon className="w-4" />;
-        case StructureTypes.Coloring:
-            return <SwatchIcon className="w-4" />;
-        case StructureTypes.Section:
-            return <></>;
-        default:
-            return <QuestionMarkCircleIcon className="w-4"/>;
-    }
 }
